Extract out itemVariationsDAO, sessionsDAO
